Minor fixes

+++ case-study-01/run-services.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-mvn clean install
-docker-compose down --remove-orphans
-docker-compose build --no-cache
-docker-compose up
+num_args=$#
+
+if [[ $num_args -eq 1 && "$1" == "--clean" ]]; then
+  ./mvnw clean install
+else
+  ./mvnw install
+fi
+
+if [[ $? -eq 0 ]]; then
+  docker-compose down --remove-orphans
+  docker-compose build --no-cache
+  docker-compose up
+fi
